+++ core-ng/src/main/java/core/framework/api/module/LogConfig.java
@@ -3,11 +3,14 @@ package core.framework.api.module;
 import core.framework.impl.log.ActionLogger;
 import core.framework.impl.log.LogForwarder;
 import core.framework.impl.log.TraceLogger;
+import core.framework.impl.log.stat.CollectStatJob;
 import core.framework.impl.module.ModuleContext;
+import core.framework.impl.scheduler.FixedRateTrigger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.nio.file.Path;
+import java.time.Duration;
 
 /**
  * @author neo
@@ -58,6 +61,7 @@ public final class LogConfig {
             logger.info("disable log forwarding during test");
         } else {
             context.logManager.logForwarder = new LogForwarder(host, context.logManager.appName);
+            context.scheduler().addTrigger(new FixedRateTrigger("collect-stat", new CollectStatJob(context.logManager.logForwarder), Duration.ofSeconds(10)));
         }
     }
 }