+++ core-ng/src/main/java/core/framework/impl/log/LogForwarder.java
@@ -3,7 +3,7 @@ package core.framework.impl.log;
 import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.ShutdownSignalException;
-import core.framework.api.queue.Message;
+import core.framework.api.util.Exceptions;
 import core.framework.api.util.JSON;
 import core.framework.api.util.Maps;
 import core.framework.api.util.Network;
@@ -11,13 +11,16 @@ import core.framework.api.util.Strings;
 import core.framework.api.util.Threads;
 import core.framework.impl.log.queue.ActionLogMessage;
 import core.framework.impl.log.queue.PerformanceStatMessage;
+import core.framework.impl.log.queue.StatMessage;
 import core.framework.impl.queue.RabbitMQ;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.time.Duration;
+import java.time.Instant;
 import java.util.Map;
+import java.util.UUID;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeoutException;
@@ -30,9 +33,9 @@ public final class LogForwarder {
     private final Logger logger = LoggerFactory.getLogger(LogForwarder.class);
     private final String appName;
 
-    private final AMQP.BasicProperties properties = new AMQP.BasicProperties.Builder().type(ActionLogMessage.class.getAnnotation(Message.class).name()).build();
+    private final AMQP.BasicProperties properties = new AMQP.BasicProperties.Builder().build();
 
-    private final BlockingQueue<ActionLogMessage> queue = new LinkedBlockingQueue<>();
+    private final BlockingQueue<Object> queue = new LinkedBlockingQueue<>();
     private final RabbitMQ rabbitMQ = new RabbitMQ();
 
     private final AtomicBoolean stop = new AtomicBoolean(false);
@@ -47,15 +50,13 @@ public final class LogForwarder {
             logger.info("log forwarder thread started");
             while (!stop.get()) {
                 try {
-                    sendActionLogs();
+                    forwardLogs();
                 } catch (Throwable e) {
                     if (!stop.get()) {  // if not initiated by shutdown, exception types can be ShutdownSignalException, InterruptedException
-
                         retryAttempts++;
                         if (retryAttempts >= 10) {  // roughly 5 mins to hold messages if log queue is not available
                             queue.clear();    // clear messages to clean up memory
                         }
-
                         logger.warn("failed to send log message, retry in 30 seconds, attempts={}", retryAttempts, e);
                         Threads.sleepRoughly(Duration.ofSeconds(30));
                     }
@@ -77,17 +78,27 @@ public final class LogForwarder {
         rabbitMQ.close();
     }
 
-    private void sendActionLogs() throws InterruptedException, IOException {
+    private void forwardLogs() throws InterruptedException, IOException {
         Channel channel = rabbitMQ.createChannel();
         try {
             while (!stop.get()) {
-                ActionLogMessage message = queue.take();
-                channel.basicPublish("", "action-log-queue", properties, Strings.bytes(JSON.toJSON(message)));
+                Object message = queue.take();
+                String queueName = queueName(message);
+                channel.basicPublish("", queueName, properties, Strings.bytes(JSON.toJSON(message)));
                 retryAttempts = 0;  // reset retry attempts if one message sent successfully
             }
         } finally {
             closeChannel(channel);
         }
+    }
+
+    private String queueName(Object message) {
+        if (message instanceof ActionLogMessage) {
+            return "action-log-queue";
+        } else if (message instanceof StatMessage) {
+            return "stat-queue";
+        }
+        throw Exceptions.error("unexpected message, messageClass={}", message.getClass().getCanonicalName());
     }
 
     private void closeChannel(Channel channel) {
@@ -100,7 +111,7 @@ public final class LogForwarder {
         }
     }
 
-    void forward(ActionLog log) {
+    void forwardLog(ActionLog log) {
         ActionLogMessage message = new ActionLogMessage();
         message.app = appName;
         message.serverIP = Network.localHostAddress();
@@ -113,7 +124,6 @@ public final class LogForwarder {
         message.errorCode = log.errorCode();
         message.errorMessage = log.errorMessage;
         message.context = log.context;
-
         Map<String, PerformanceStatMessage> performanceStats = Maps.newLinkedHashMap();
         log.performanceStats.forEach((key, stat) -> {
             PerformanceStatMessage statMessage = new PerformanceStatMessage();
@@ -122,7 +132,6 @@ public final class LogForwarder {
             performanceStats.put(key, statMessage);
         });
         message.performanceStats = performanceStats;
-
         if (log.flushTraceLog()) {
             StringBuilder traceLog = new StringBuilder(log.events.size() * 64);
             for (LogEvent event : log.events) {
@@ -130,7 +139,16 @@ public final class LogForwarder {
             }
             message.traceLog = traceLog.toString();
         }
+        queue.add(message);
+    }
 
+    public void forwardStats(Map<String, Double> stats) {
+        StatMessage message = new StatMessage();
+        message.id = UUID.randomUUID().toString();
+        message.date = Instant.now();
+        message.app = appName;
+        message.serverIP = Network.localHostAddress();
+        message.stats = stats;
         queue.add(message);
     }
 }