+++ core-ng/src/main/java/core/framework/impl/module/ModuleContext.java
@@ -66,10 +66,10 @@ public final class ModuleContext {
 
         if (!isTest()) {
             httpServer.handler.route.add(HTTPMethod.GET, "/health-check", new ControllerHolder(new HealthCheckController(), true));
-            httpServer.handler.route.add(HTTPMethod.GET, "/monitor/memory", new ControllerHolder(new MemoryUsageController(), true));
+            httpServer.handler.route.add(HTTPMethod.GET, "/management/memory", new ControllerHolder(new MemoryUsageController(), true));
             ThreadInfoController threadInfoController = new ThreadInfoController();
-            httpServer.handler.route.add(HTTPMethod.GET, "/monitor/thread", new ControllerHolder(threadInfoController::threadUsage, true));
-            httpServer.handler.route.add(HTTPMethod.GET, "/monitor/thread-dump", new ControllerHolder(threadInfoController::threadDump, true));
+            httpServer.handler.route.add(HTTPMethod.GET, "/management/thread", new ControllerHolder(threadInfoController::threadUsage, true));
+            httpServer.handler.route.add(HTTPMethod.GET, "/management/thread-dump", new ControllerHolder(threadInfoController::threadDump, true));
         }
     }